Updated files to separate modules for img/video
New modules for logo and video detection. Updated readme
