+++ README.md
@@ -1,14 +1,14 @@
-# Streamlit App for Blood Cell Count Dataset
+# Streamlit App for Face Detection Dataset (by Mohamed Traore)
 
-This app allows you to upload an image of blood cells and use an [object detection model](https://models.roboflow.com/object-detection) to detect the white blood cells, red blood cells, and platelets.
+This app allows you to upload an image of faces and use an [object detection model](https://models.roboflow.com/object-detection) to detect faces in an image, and facdes in a video feed.
 
-The app is [found here](https://roboflow.com/streamlit-bccd):
+The app is [found here](https://share.streamlit.io/motraor3/streamlit-bccd):
 ```
-https://roboflow.com/streamlit-bccd
+https://share.streamlit.io/motraor3/streamlit-bccd
 ```
 
-The app will work as is. If you want to build your own model, you'll need your own API key. [Create a Roboflow account](https://app.roboflow.com) to get your own API key.
+The app will work as is. If you want to build your own model, you'll need your own API key. [Create a Roboflow account](https://app.roboflow.com) to get your own API key. [Obtaining your API key](https://docs.roboflow.com/rest-api#obtaining-your-api-key)
 
-Want to [build and deploy your own object detection model and app](https://blog.streamlit.io/how-to-use-roboflow-and-streamlit-to-visualize-object-detection-output/)? Follow the steps in our "How To" blog!
+Want to build and deploy your own object detection model and app? [Follow this link](https://blog.streamlit.io/how-to-use-roboflow-and-streamlit-to-visualize-object-detection-output/) for how to upload images for inference, and [this tutorial](https://www.youtube.com/watch?v=w4fgZg-jb28) shows how to add a video feed for inference.
 
 This app was created using [Roboflow](https://roboflow.com) and [Streamlit](https://streamlit.io/).
